pallet_revive_uapi::precompiles::system

Module ISystem

Module containing a contract’s types and functions.

interface ISystem {
    function hashBlake256(bytes memory input) external pure returns (bytes32 digest);
    function hashBlake128(bytes memory input) external pure returns (bytes32 digest);
    function toAccountId(address input) external view returns (bytes memory account_id);
    function callerIsOrigin() external view returns (bool);
    function callerIsRoot() external view returns (bool);
    function minimumBalance() external view returns (uint);
    function ownCodeHash() external view returns (bytes32);
    function weightLeft() external view returns (uint64 refTime, uint64 proofSize);
    function terminate(address beneficiary) external;
}

Structs

callerIsOriginCall

Checks whether the caller of the contract calling this function is the origin of the whole call stack. Function with signature callerIsOrigin() and selector 0x9098d8c1.

callerIsOriginReturn

Checks whether the caller of the contract calling this function is the origin of the whole call stack. Container type for the return parameters of the callerIsOrigin() function.

callerIsRootCall

Checks whether the caller of the contract calling this function is root.

callerIsRootReturn

Checks whether the caller of the contract calling this function is root.

hashBlake128Call

Computes the BLAKE2 128-bit hash on the given input. Function with signature hashBlake128(bytes) and selector 0xa63b6983.

hashBlake128Return

Computes the BLAKE2 128-bit hash on the given input. Container type for the return parameters of the hashBlake128(bytes) function.

hashBlake256Call

Computes the BLAKE2 256-bit hash on the given input. Function with signature hashBlake256(bytes) and selector 0xc7dff686.

hashBlake256Return

Computes the BLAKE2 256-bit hash on the given input. Container type for the return parameters of the hashBlake256(bytes) function.

minimumBalanceCall

Returns the minimum balance that is required for creating an account (the existential deposit). Function with signature minimumBalance() and selector 0xb9d1d49b.

minimumBalanceReturn

Returns the minimum balance that is required for creating an account (the existential deposit). Container type for the return parameters of the minimumBalance() function.

ownCodeHashCall

Returns the code hash of the caller. Function with signature ownCodeHash() and selector 0xdb6bdc8a.

ownCodeHashReturn

Returns the code hash of the caller. Container type for the return parameters of the ownCodeHash() function.

terminateCall

Terminate the calling contract of this function and send balance to beneficiary. This will revert if:

terminateReturn

Terminate the calling contract of this function and send balance to beneficiary. This will revert if:

toAccountIdCall

Retrieve the account id for a specified H160 address.

toAccountIdReturn

Retrieve the account id for a specified H160 address.

weightLeftCall

Returns the amount of Weight left. Function with signature weightLeft() and selector 0x8d29dffb.

weightLeftReturn

Returns the amount of Weight left. Container type for the return parameters of the weightLeft() function.

Enums

ISystemCalls

Container for all the ISystem function calls.